Compile Ruby with only a single job.
The compilation appears to be flaky.
In my limited testing, I noticed that an "mkdir" command tried to create a directory at "$buildroot/ruby-1.9.3-p551/.ext/common/digest", but this was already a file containing Ruby code.
It may be a temporary file with the unfortunate side-effect of concurrency issues when building.

--HG--
branch : updates